Create class for storage

+++ src/utils/storage.py
@@ -2,41 +2,32 @@ from azure.storage.blob import BlobServiceClient
 from config import get_settings
 
 
-def config():
-    """
-    Get the values of env and create container client
-    """
-    # TODO: Considerar si este script deberia ser una clase
-    settings = get_settings()
+class Storage:
 
-    blob_service = BlobServiceClient.from_connection_string(
-        settings.abs_string
-    )
-    container_client = blob_service.get_container_client(
-        settings.abs_container
-    )
-
-    return container_client
-
-
-def upload_file(file_name, file_path):
-    """
-    Upload a blob to Azure Blob Storage
-    """
-    container_client = config()
-
-    with open(file_path, "rb") as data:
-        blob_client = container_client.upload_blob(
-            file_name, data, overwrite=True
+    def __init__(self):
+        settings = get_settings()
+        blob_service = BlobServiceClient.from_connection_string(
+            settings.abs_string
         )
-        return blob_client.url
+        self.container_client = blob_service.get_container_client(
+            settings.abs_container
+        )
 
+    def upload_file(self, file_name, file_path):
+        """
+        Upload a blob to Azure Blob Storage
+        """
 
-def remove_file(file_path):
-    """
-    Remove a blob from Azure Blob Storage
-    """
-    container_client = config()
+        with open(file_path, "rb") as data:
+            blob_client = self.container_client.upload_blob(
+                file_name, data, overwrite=True
+            )
+            return blob_client.url
 
-    blob_client = container_client.get_blob_client(file_path)
-    blob_client.delete_blob()
+    def remove_file(self, file_path):
+        """
+        Remove a blob from Azure Blob Storage
+        """
+        blob_client = self.container_client.get_blob_client(file_path)
+        if blob_client:
+            blob_client.delete_blob()
